:hammer: refactor: Reuse duplicated code to modify
