+++ types/channel/socket/member-info.interface.ts
@@ -1,8 +1,8 @@
-export type Role = 'owner' | 'admin' | 'member';
+export type ChannelRole = 'owner' | 'admin' | 'member';
 
 export interface MemberInfo {
   userId: number;
   nickname: string;
   image?: string;
-  role: Role;
+  role: ChannelRole;
 }
